LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-element-036-expected.txt


FAIL display:block img should be 200px high assert_equals: expected 200 but got 100
FAIL display:inline-block img should be 200px high assert_equals: expected 200 but got 100
